finished DI example
