+++ pom.xml
@@ -47,6 +47,11 @@
             <version>4.12</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>RELEASE</version>
+        </dependency>
         <!--<dependency>-->
     		<!--<groupId>org.picocontainer</groupId>  -->
     		<!--<artifactId>picocontainer</artifactId>-->
