--- src/main/java/com/cucumber/helpers/Helpers.java
@@ -1,5 +1,0 @@
-package com.cucumber.helpers;
-
-public class Helpers {
-
-}
