--- src/test/java/com/cucumber/steps/Hooks.java
@@ -1,30 +1,0 @@
-package com.cucumber.steps;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-import cucumber.api.java.After;
-import cucumber.api.java.Before;
-
-import cucumber.api.Scenario;
-
-//import cucumber.api.picocontainer;
-
-import org.picocontainer.DefaultPicoContainer;
-
-public class Hooks {
-	
-	private WebDriver driver;
-
-	@Before
-	public void beforeScenario(Scenario scenario) {
-
-
-	}
-
-	
-	@After
-	public void afterScenario(Scenario scenario) {
-
-	}
-}
