+++ src/test/java/cucumber/support/RunCukeTest.java
@@ -1,14 +1,15 @@
-package com.cucumber.steps;
+package cucumber.support;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
-
+//(features="src/test/resources")
 @RunWith(Cucumber.class)
 @CucumberOptions(features = "classpath:features",
-				 format = { "pretty",
-						 	"html:target/cucumber-html-report" },
-				 tags = {})
+		plugin={"pretty"},
+		glue = {"cucumber.steps",
+                "cucumber.support"},
+		tags = {})
 public class RunCukeTest {
